test/com/assignment4/SystemFacadeTest.java
package com.assignment4;

import com.assignment4.facade.SystemFacade;
import org.junit.Test;

import static org.junit.Assert.*;

public class SystemFacadeTest {
    @Test
    public void testSuccessfulDatabaseOperation() {
        SystemFacade facade = new SystemFacade();
        boolean result = facade.performDatabaseOperation("valid-credentials", "SELECT * FROM users");
        assertTrue("Database operation should succeed with valid credentials", result);
    }

    @Test
    public void testFacadeEncapsulation() {
        SystemFacade facade = new SystemFacade();
        assertNotNull("Facade should be properly initialized", facade);
    }
    @Test
    public void testComplexOperationSimplification() {
        SystemFacade facade = new SystemFacade();
        boolean result = facade.performDatabaseOperation("valid-credentials", "INSERT INTO logs VALUES(1)");
        assertTrue("Complex operation should be simplified through facade", result);
    }
    }

